backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# NLP
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
twikit==1.6.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
